+++ fizzim.pl
@@ -2970,6 +2970,7 @@ sub merge_fork_transition {
   }
   if (exists $outref->{attributes}) {
     foreach $att (keys %{ $outref->{attributes} }) {
+      next if (&preserve_incoming_fork_action($inref, $outref, $att));
       if (ref($outref->{attributes}{$att}) eq "HASH") {
         $merged{attributes}{$att} = { %{ $outref->{attributes}{$att} } };
       } else {
@@ -2991,15 +2992,39 @@ sub merge_fork_transition {
   return %merged;
 }
 
+sub preserve_incoming_fork_action {
+  my ($inref, $outref, $att) = @_;
+  my ($in_value, $out_value, $out_type);
+
+  return 0 unless exists $inref->{attributes}{$att};
+  return 0 unless exists $outref->{attributes}{$att};
+  $out_type = $outref->{attributes}{$att}{type};
+  return 0 unless defined $out_type && $out_type eq "output";
+
+  $in_value = $inref->{attributes}{$att}{value};
+  $out_value = $outref->{attributes}{$att}{value};
+  $in_value =~ s/^\s+|\s+$//g if defined $in_value;
+  $out_value =~ s/^\s+|\s+$//g if defined $out_value;
+
+  return (defined $in_value && $in_value ne "" && (!defined $out_value || $out_value eq ""));
+}
+
 sub combine_fork_priorities {
   my ($in_priority, $out_priority, $out_index, $out_count) = @_;
+  my ($int_part, $frac_part, $branch_rank);
 
   $in_priority =~ s/^\s+|\s+$//g if defined $in_priority;
   $out_priority =~ s/^\s+|\s+$//g if defined $out_priority;
 
   if (defined $in_priority && ($in_priority ne "")) {
     return $in_priority if (!defined $out_count || $out_count <= 1);
-    return sprintf("%.9f", $in_priority + (($out_index + 1) / 1000000000));
+    $branch_rank = sprintf("%03d", $out_index + 1);
+    if ($in_priority =~ /^(-?\d+)(?:\.(\d+))?$/) {
+      $int_part = $1;
+      $frac_part = defined $2 ? $2 : "";
+      return "$int_part.$frac_part$branch_rank";
+    }
+    return $in_priority;
   }
   return $out_priority if (defined $out_priority);
   return "";